@@ -7,36 +7,30 @@ classifiers =
 	License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)
 	Operating System :: OS Independent
 	Programming Language :: Python
-	Programming Language :: Python :: 2
-	Programming Language :: Python :: 2.7
 	Programming Language :: Python :: 3
-	Programming Language :: Python :: 3.5
-	Programming Language :: Python :: 3.6
 	Programming Language :: Python :: 3.7
 	Programming Language :: Python :: 3.8
 	Programming Language :: Python :: 3.9
+	Programming Language :: Python :: 3.10
+	Programming Language :: Python :: 3.11
 	Programming Language :: Python :: Implementation :: CPython
 	Programming Language :: Python :: Implementation :: PyPy
 	Topic :: Software Development :: Libraries :: Python Modules
 	Topic :: Utilities
 description = Utility library for gitignore style pattern matching of file paths.
 license = MPL 2.0
-long_description = file: README.rst, CHANGES.rst
+long_description = file: README-dist.rst
 long_description_content_type = text/x-rst
 name = pathspec
-version = attr: pathspec_meta.__version__
-url = https://github.com/cpburnz/python-path-specification
+url = https://github.com/cpburnz/python-pathspec
+version = attr: pathspec._meta.__version__
 
 [options]
 packages = find:
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*
-setup_requires = setuptools >=39.2.0
-test_suite = pathspec.tests
+python_requires = >=3.7
+setup_requires = setuptools>=40.8.0
+test_suite = tests
 
-[bdist_wheel]
-universal = 1
-
-[egg_info]
-tag_build = 
-tag_date = 0
+[options.packages.find]
+include = pathspec, pathspec.*
 
