@@ -1,23 +1,20 @@
-# encoding: utf-8
 """
 This module provides the base definition for patterns.
 """
 
+import dataclasses
 import re
-try:
-	from typing import (
-		AnyStr,
-		Iterable,
-		Iterator,
-		Optional,
-		Pattern as RegexHint,
-		Text,
-		Tuple,
-		Union)
-except ImportError:
-	pass
-
-from .compat import unicode
+import warnings
+from typing import (
+	Any,
+	AnyStr,
+	Iterable,
+	Iterator,
+	Match as MatchHint,
+	Optional,
+	Pattern as PatternHint,
+	Tuple,
+	Union)
 
 
 class Pattern(object):
@@ -28,8 +25,7 @@ class Pattern(object):
 	# Make the class dict-less.
 	__slots__ = ('include',)
 
-	def __init__(self, include):
-		# type: (Optional[bool]) -> None
+	def __init__(self, include: Optional[bool]) -> None:
 		"""
 		Initializes the :class:`Pattern` instance.
 
@@ -45,18 +41,42 @@ class Pattern(object):
 		or is a null-operation (:data:`None`).
 		"""
 
-	def match(self, files):
-		# type: (Iterable[Text]) -> Iterator[Text]
+	def match(self, files: Iterable[str]) -> Iterator[str]:
 		"""
+		DEPRECATED: This method is no longer used and has been replaced by
+		:meth:`.match_file`. Use the :meth:`.match_file` method with a loop
+		for similar results.
+
 		Matches this pattern against the specified files.
 
-		*files* (:class:`~collections.abc.Iterable` of :class:`str`) contains
-		each file relative to the root directory (e.g., ``"relative/path/to/file"``).
+		*files* (:class:`~collections.abc.Iterable` of :class:`str`)
+		contains each file relative to the root directory (e.g.,
+		:data:`"relative/path/to/file"`).
 
 		Returns an :class:`~collections.abc.Iterable` yielding each matched
 		file path (:class:`str`).
 		"""
-		raise NotImplementedError("{}.{} must override match().".format(self.__class__.__module__, self.__class__.__name__))
+		warnings.warn((
+			"{0.__module__}.{0.__qualname__}.match() is deprecated. Use "
+			"{0.__module__}.{0.__qualname__}.match_file() with a loop for "
+			"similar results."
+		).format(self.__class__), DeprecationWarning, stacklevel=2)
+
+		for file in files:
+			if self.match_file(file) is not None:
+				yield file
+
+	def match_file(self, file: str) -> Optional[Any]:
+		"""
+		Matches this pattern against the specified file.
+
+		*file* (:class:`str`) is the normalized file path to match against.
+
+		Returns the match result if *file* matched; otherwise, :data:`None`.
+		"""
+		raise NotImplementedError((
+			"{0.__module__}.{0.__qualname__} must override match_file()."
+		).format(self.__class__))
 
 
 class RegexPattern(Pattern):
@@ -65,20 +85,22 @@ class RegexPattern(Pattern):
 	using regular expressions.
 	"""
 
-	# Make the class dict-less.
+	# Keep the class dict-less.
 	__slots__ = ('regex',)
 
-	def __init__(self, pattern, include=None):
-		# type: (Union[AnyStr, RegexHint], Optional[bool]) -> None
+	def __init__(
+		self,
+		pattern: Union[AnyStr, PatternHint],
+		include: Optional[bool] = None,
+	) -> None:
 		"""
 		Initializes the :class:`RegexPattern` instance.
 
-		*pattern* (:class:`unicode`, :class:`bytes`, :class:`re.RegexObject`,
-		or :data:`None`) is the pattern to compile into a regular
-		expression.
+		*pattern* (:class:`str`, :class:`bytes`, :class:`re.Pattern`, or
+		:data:`None`) is the pattern to compile into a regular expression.
 
 		*include* (:class:`bool` or :data:`None`) must be :data:`None`
-		unless *pattern* is a precompiled regular expression (:class:`re.RegexObject`)
+		unless *pattern* is a precompiled regular expression (:class:`re.Pattern`)
 		in which case it is whether matched files should be included
 		(:data:`True`), excluded (:data:`False`), or is a null operation
 		(:data:`None`).
@@ -87,14 +109,10 @@ class RegexPattern(Pattern):
 			   parameter.
 		"""
 
-		self.regex = None
-		"""
-		*regex* (:class:`re.RegexObject`) is the regular expression for the
-		pattern.
-		"""
-
-		if isinstance(pattern, (unicode, bytes)):
-			assert include is None, "include:{!r} must be null when pattern:{!r} is a string.".format(include, pattern)
+		if isinstance(pattern, (str, bytes)):
+			assert include is None, (
+				"include:{!r} must be null when pattern:{!r} is a string."
+			).format(include, pattern)
 			regex, include = self.pattern_to_regex(pattern)
 			# NOTE: Make sure to allow a null regular expression to be
 			# returned for a null-operation.
@@ -109,16 +127,22 @@ class RegexPattern(Pattern):
 		elif pattern is None:
 			# NOTE: Make sure to allow a null pattern to be passed for a
 			# null-operation.
-			assert include is None, "include:{!r} must be null when pattern:{!r} is null.".format(include, pattern)
+			assert include is None, (
+				"include:{!r} must be null when pattern:{!r} is null."
+			).format(include, pattern)
 
 		else:
-			raise TypeError("pattern:{!r} is not a string, RegexObject, or None.".format(pattern))
+			raise TypeError("pattern:{!r} is not a string, re.Pattern, or None.".format(pattern))
 
 		super(RegexPattern, self).__init__(include)
-		self.regex = regex
 
-	def __eq__(self, other):
-		# type: (RegexPattern) -> bool
+		self.regex: PatternHint = regex
+		"""
+		*regex* (:class:`re.Pattern`) is the regular expression for the
+		pattern.
+		"""
+
+	def __eq__(self, other: 'RegexPattern') -> bool:
 		"""
 		Tests the equality of this regex pattern with *other* (:class:`RegexPattern`)
 		by comparing their :attr:`~Pattern.include` and :attr:`~RegexPattern.regex`
@@ -129,25 +153,25 @@ class RegexPattern(Pattern):
 		else:
 			return NotImplemented
 
-	def match(self, files):
-		# type: (Iterable[Text]) -> Iterable[Text]
+	def match_file(self, file: str) -> Optional['RegexMatchResult']:
 		"""
-		Matches this pattern against the specified files.
+		Matches this pattern against the specified file.
 
-		*files* (:class:`~collections.abc.Iterable` of :class:`str`)
+		*file* (:class:`str`)
 		contains each file relative to the root directory (e.g., "relative/path/to/file").
 
-		Returns an :class:`~collections.abc.Iterable` yielding each matched
-		file path (:class:`str`).
+		Returns the match result (:class:`RegexMatchResult`) if *file*
+		matched; otherwise, :data:`None`.
 		"""
 		if self.include is not None:
-			for path in files:
-				if self.regex.match(path) is not None:
-					yield path
+			match = self.regex.match(file)
+			if match is not None:
+				return RegexMatchResult(match)
+
+		return None
 
 	@classmethod
-	def pattern_to_regex(cls, pattern):
-		# type: (Text) -> Tuple[Text, bool]
+	def pattern_to_regex(cls, pattern: str) -> Tuple[str, bool]:
 		"""
 		Convert the pattern into an uncompiled regular expression.
 
@@ -162,3 +186,21 @@ class RegexPattern(Pattern):
 			   :data:`True`.
 		"""
 		return pattern, True
+
+
+@dataclasses.dataclass()
+class RegexMatchResult(object):
+	"""
+	The :class:`RegexMatchResult` data class is used to return information
+	about the matched regular expression.
+	"""
+
+	# Keep the class dict-less.
+	__slots__ = (
+		'match',
+	)
+
+	match: MatchHint
+	"""
+	*match* (:class:`re.Match`) is the regex match result.
+	"""
