@@ -1,4 +1,3 @@
-# encoding: utf-8
 """
 The *pathspec* package provides pattern matching for file paths. So far
 this only includes Git's wildmatch pattern matching (the style used for
@@ -7,25 +6,41 @@ this only includes Git's wildmatch pattern matching (the style used for
 The following classes are imported and made available from the root of
 the `pathspec` package:
 
-- :class:`pathspec.pathspec.PathSpec`
+-	:class:`pathspec.gitignore.GitIgnoreSpec`
 
-- :class:`pathspec.pattern.Pattern`
+-	:class:`pathspec.pathspec.PathSpec`
 
-- :class:`pathspec.pattern.RegexPattern`
+-	:class:`pathspec.pattern.Pattern`
 
-- :class:`pathspec.util.RecursionError`
+-	:class:`pathspec.pattern.RegexPattern`
+
+-	:class:`pathspec.util.RecursionError`
 
 The following functions are also imported:
 
-- :func:`pathspec.util.iter_tree`
-- :func:`pathspec.util.lookup_pattern`
-- :func:`pathspec.util.match_files`
-"""
-from __future__ import unicode_literals
+-	:func:`pathspec.util.lookup_pattern`
 
-from .pathspec import PathSpec
-from .pattern import Pattern, RegexPattern
-from .util import iter_tree, lookup_pattern, match_files, RecursionError
+The following deprecated functions are also imported to maintain
+backward compatibility:
+
+-	:func:`pathspec.util.iter_tree` which is an alias for
+	:func:`pathspec.util.iter_tree_files`.
+
+-	:func:`pathspec.util.match_files`
+"""
+
+from .gitignore import (
+	GitIgnoreSpec)
+from .pathspec import (
+	PathSpec)
+from .pattern import (
+	Pattern,
+	RegexPattern)
+from .util import (
+	RecursionError,
+	iter_tree,
+	lookup_pattern,
+	match_files)
 
 from ._meta import (
 	__author__,
@@ -38,6 +53,24 @@ from ._meta import (
 # Load pattern implementations.
 from . import patterns
 
-# Expose `GitIgnorePattern` class in the root module for backward
-# compatibility with v0.4.
+# DEPRECATED: Expose the `GitIgnorePattern` class in the root module for
+# backward compatibility with v0.4.
 from .patterns.gitwildmatch import GitIgnorePattern
+
+# Declare private imports as part of the public interface. Deprecated
+# imports are deliberately excluded.
+__all__ = [
+	'GitIgnoreSpec',
+	'PathSpec',
+	'Pattern',
+	'RecursionError',
+	'RegexPattern',
+	'__author__',
+	'__copyright__',
+	'__credits__',
+	'__license__',
+	'__version__',
+	'iter_tree',
+	'lookup_pattern',
+	'match_files',
+]
