@@ -1,4 +1,3 @@
-# encoding: utf-8
 """
 This script tests utility functions.
 """
@@ -6,56 +5,87 @@ This script tests utility functions.
 import errno
 import os
 import os.path
+import pathlib
 import shutil
-import sys
 import tempfile
 import unittest
+from functools import (
+	partial)
+from typing import (
+	Iterable,
+	Tuple)
 
-from pathspec.util import iter_tree_entries, iter_tree_files, RecursionError, normalize_file
+from pathspec.patterns.gitwildmatch import (
+	GitWildMatchPattern)
+from pathspec.util import (
+	RecursionError,
+	iter_tree_entries,
+	iter_tree_files,
+	match_file,
+	normalize_file)
+from tests.util import (
+	make_dirs,
+	make_files,
+	make_links,
+	mkfile,
+	ospath)
+
+
+class MatchFileTest(unittest.TestCase):
+	"""
+	The :class:`MatchFileTest` class tests the :meth:`.match_file`
+	function.
+	"""
+
+	def test_1_match_file(self):
+		"""
+		Test matching files individually.
+		"""
+		patterns = list(map(GitWildMatchPattern, [
+			'*.txt',
+			'!b.txt',
+		]))
+		results = set(filter(partial(match_file, patterns), [
+			'X/a.txt',
+			'X/b.txt',
+			'X/Z/c.txt',
+			'Y/a.txt',
+			'Y/b.txt',
+			'Y/Z/c.txt',
+		]))
+		self.assertEqual(results, {
+			'X/a.txt',
+			'X/Z/c.txt',
+			'Y/a.txt',
+			'Y/Z/c.txt',
+		})
 
 
 class IterTreeTest(unittest.TestCase):
 	"""
-	The ``IterTreeTest`` class tests `pathspec.util.iter_tree_files()`.
+	The :class:`IterTreeTest` class tests :meth:`.iter_tree_entries` and
+	:meth:`.iter_tree_files` functions.
 	"""
 
-	def make_dirs(self, dirs):
+	def make_dirs(self, dirs: Iterable[str]) -> None:
 		"""
 		Create the specified directories.
 		"""
-		for dir in dirs:
-			os.mkdir(os.path.join(self.temp_dir, self.ospath(dir)))
+		make_dirs(self.temp_dir, dirs)
 
-	def make_files(self, files):
+	def make_files(self, files: Iterable[str]) -> None:
 		"""
 		Create the specified files.
 		"""
-		for file in files:
-			self.mkfile(os.path.join(self.temp_dir, self.ospath(file)))
+		make_files(self.temp_dir, files)
 
-	def make_links(self, links):
+	def make_links(self, links: Iterable[Tuple[str, str]]) -> None:
 		"""
 		Create the specified links.
 		"""
-		for link, node in links:
-			os.symlink(os.path.join(self.temp_dir, self.ospath(node)), os.path.join(self.temp_dir, self.ospath(link)))
+		make_links(self.temp_dir, links)
 
-	@staticmethod
-	def mkfile(file):
-		"""
-		Creates an empty file.
-		"""
-		with open(file, 'wb'):
-			pass
-
-	@staticmethod
-	def ospath(path):
-		"""
-		Convert the POSIX path to a native OS path.
-		"""
-		return os.path.join(*path.split('/'))
-
-	def require_realpath(self):
+	def require_realpath(self) -> None:
 		"""
 		Skips the test if `os.path.realpath` does not properly support
 		symlinks.
@@ -63,20 +93,20 @@ class IterTreeTest(unittest.TestCase):
 		if self.broken_realpath:
 			raise unittest.SkipTest("`os.path.realpath` is broken.")
 
-	def require_symlink(self):
+	def require_symlink(self) -> None:
 		"""
 		Skips the test if `os.symlink` is not supported.
 		"""
 		if self.no_symlink:
 			raise unittest.SkipTest("`os.symlink` is not supported.")
 
-	def setUp(self):
+	def setUp(self) -> None:
 		"""
 		Called before each test.
 		"""
-		self.temp_dir = tempfile.mkdtemp()
+		self.temp_dir = pathlib.Path(tempfile.mkdtemp())
 
-	def tearDown(self):
+	def tearDown(self) -> None:
 		"""
 		Called after each test.
 		"""
@@ -100,7 +130,7 @@ class IterTreeTest(unittest.TestCase):
 			'Dir/Inner/f',
 		])
 		results = set(iter_tree_files(self.temp_dir))
-		self.assertEqual(results, set(map(self.ospath, [
+		self.assertEqual(results, set(map(ospath, [
 			'a',
 			'b',
 			'Dir/c',
@@ -113,20 +143,20 @@ class IterTreeTest(unittest.TestCase):
 		"""
 		Tests whether links can be created.
 		"""
-		# NOTE: Windows does not support `os.symlink` for Python 2. Windows
-		# Vista and greater supports `os.symlink` for Python 3.2+.
+		# NOTE: Windows Vista and greater supports `os.symlink` for Python
+		# 3.2+.
 		no_symlink = None
 		try:
-			file = os.path.join(self.temp_dir, 'file')
-			link = os.path.join(self.temp_dir, 'link')
-			self.mkfile(file)
+			file = self.temp_dir / 'file'
+			link = self.temp_dir / 'link'
+			mkfile(file)
 
 			try:
 				os.symlink(file, link)
-			except (AttributeError, NotImplementedError):
+			except (AttributeError, NotImplementedError, OSError):
 				no_symlink = True
-				raise
-			no_symlink = False
+			else:
+				no_symlink = False
 
 		finally:
 			self.__class__.no_symlink = no_symlink
@@ -141,17 +171,17 @@ class IterTreeTest(unittest.TestCase):
 		broken_realpath = None
 		try:
 			self.require_symlink()
-			file = os.path.join(self.temp_dir, 'file')
-			link = os.path.join(self.temp_dir, 'link')
-			self.mkfile(file)
+			file = self.temp_dir / 'file'
+			link = self.temp_dir / 'link'
+			mkfile(file)
 			os.symlink(file, link)
 
 			try:
 				self.assertEqual(os.path.realpath(file), os.path.realpath(link))
 			except AssertionError:
 				broken_realpath = True
-				raise
-			broken_realpath = False
+			else:
+				broken_realpath = False
 
 		finally:
 			self.__class__.broken_realpath = broken_realpath
@@ -178,7 +208,7 @@ class IterTreeTest(unittest.TestCase):
 			('DirX', 'Dir'),
 		])
 		results = set(iter_tree_files(self.temp_dir))
-		self.assertEqual(results, set(map(self.ospath, [
+		self.assertEqual(results, set(map(ospath, [
 			'a',
 			'ax',
 			'b',
@@ -215,7 +245,7 @@ class IterTreeTest(unittest.TestCase):
 			('Dir/Fx', 'Dir/Target'),
 		])
 		results = set(iter_tree_files(self.temp_dir))
-		self.assertEqual(results, set(map(self.ospath, [
+		self.assertEqual(results, set(map(ospath, [
 			'Ax/Ex/file',
 			'Ax/Fx/file',
 			'Ax/Target/file',
@@ -246,8 +276,9 @@ class IterTreeTest(unittest.TestCase):
 		])
 		with self.assertRaises(RecursionError) as context:
 			set(iter_tree_files(self.temp_dir))
+
 		self.assertEqual(context.exception.first_path, 'Dir')
-		self.assertEqual(context.exception.second_path, self.ospath('Dir/Self'))
+		self.assertEqual(context.exception.second_path, ospath('Dir/Self'))
 
 	def test_2_5_recursive_circular_links(self):
 		"""
@@ -272,11 +303,12 @@ class IterTreeTest(unittest.TestCase):
 		])
 		with self.assertRaises(RecursionError) as context:
 			set(iter_tree_files(self.temp_dir))
+
 		self.assertIn(context.exception.first_path, ('A', 'B', 'C'))
 		self.assertEqual(context.exception.second_path, {
-			'A': self.ospath('A/Bx/Cx/Ax'),
-			'B': self.ospath('B/Cx/Ax/Bx'),
-			'C': self.ospath('C/Ax/Bx/Cx'),
+			'A': ospath('A/Bx/Cx/Ax'),
+			'B': ospath('B/Cx/Ax/Bx'),
+			'C': ospath('C/Ax/Bx/Cx'),
 		}[context.exception.first_path])
 
 	def test_2_6_detect_broken_links(self):
@@ -292,6 +324,7 @@ class IterTreeTest(unittest.TestCase):
 		])
 		with self.assertRaises(OSError) as context:
 			set(iter_tree_files(self.temp_dir, on_error=reraise))
+
 		self.assertEqual(context.exception.errno, errno.ENOENT)
 
 	def test_2_7_ignore_broken_links(self):
@@ -327,7 +360,7 @@ class IterTreeTest(unittest.TestCase):
 			('DirX', 'Dir'),
 		])
 		results = set(iter_tree_files(self.temp_dir, follow_links=False))
-		self.assertEqual(results, set(map(self.ospath, [
+		self.assertEqual(results, set(map(ospath, [
 			'A',
 			'Ax',
 			'B',
@@ -357,7 +390,7 @@ class IterTreeTest(unittest.TestCase):
 			'Dir/Inner/f',
 		])
 		results = {entry.path for entry in iter_tree_entries(self.temp_dir)}
-		self.assertEqual(results, set(map(self.ospath, [
+		self.assertEqual(results, set(map(ospath, [
 			'a',
 			'b',
 			'Dir',
@@ -369,12 +402,10 @@ class IterTreeTest(unittest.TestCase):
 			'Empty',
 		])))
 
-	@unittest.skipIf(sys.version_info < (3, 4), "pathlib entered stdlib in Python 3.4")
 	def test_4_normalizing_pathlib_path(self):
 		"""
-		Tests passing pathlib.Path as argument.
+		Tests normalizing a :class:`pathlib.PurePath` as argument.
 		"""
-		from pathlib import Path
-		first_spec = normalize_file(Path('a.txt'))
+		first_spec = normalize_file(pathlib.PurePath('a.txt'))
 		second_spec = normalize_file('a.txt')
 		self.assertEqual(first_spec, second_spec)
